# docs/swagger.yaml
basePath: /
definitions:
...
  mygram-byferdiansyah_socialmedia_utils.ResponseMessage:
    properties:
      data:
        example: the error explained here
        type: string
      status:
        example: fail
        type: string
    type: object
  mygram-byferdiansyah_socialmedia_utils.User:
    properties:
      email:
        example: johndoe@example.com
        type: string
      id:
        example: here is the generated user id
        type: string
      username:
        example: johndoe
        type: string
    type: object
  mygram-byferdiansyah_user_utils.ResponseMessage:
    properties:
      data:
        example: the error explained here
        type: string
      status:
        example: fail
        type: string
    type: object
...
  utils.AddSocialMedia:
    properties:
      name:
        example: Example
        type: string
      social_media_url:
        example: https://www.example.com/johndoe
        type: string
    type: object
...
  utils.AddedSocialMedia:
    properties:
      created_at:
        example: the created at generated here
        type: string
      id:
        example: the social media id generated here
        type: string
      name:
        example: Example
        type: string
      social_media_url:
        example: https://www.example.com/johndoe
        type: string
      user_id:
        example: here is the generated user id
        type: string
    type: object
...
  utils.EditSocialMedia:
    properties:
      name:
        example: New Example
        type: string
      social_media_url:
        example: https://www.newexample.com/johndoe
        type: string
    type: object
  utils.EditUser:
    properties:
      email:
        example: newjohndoe@example.com
        type: string
      username:
        example: newjohndoe
        type: string
    type: object
...
  utils.EditedSocialMedia:
    properties:
      id:
        example: here is the generated social media id
        type: string
      name:
        example: New Example
        type: string
      social_media_url:
        example: https://www.newexample.com/johndoe
        type: string
      updated_at:
        example: the edited at generated here
        type: string
      user_id:
        example: here is the generated user id
        type: string
    type: object
  utils.EditedUser:
    properties:
      age:
        example: 8
        type: integer
      email:
        example: newjohndoe@example.com
        type: string
      id:
        example: here is the generated user id
        type: string
      updated_at:
        example: the edited at generated here
        type: string
      username:
        example: newjohndoe
        type: string
    type: object
...
  utils.LoggedinUser:
    properties:
      token:
        example: the token generated here
        type: string
    type: object
  utils.LoginUser:
    properties:
      email:
        example: johndoe@example.com
        type: string
      password:
        example: secret
        type: string
    type: object
  utils.RegisterUser:
    properties:
      age:
        example: 8
        type: integer
      email:
        example: johndoe@example.com
        type: string
      password:
        example: secret
        type: string
      username:
        example: johndoe
        type: string
    type: object
  utils.RegisteredUser:
    properties:
      age:
        example: 8
        type: integer
      email:
        example: johndoe@example.com
        type: string
      id:
        example: the user id generated here
        type: string
      username:
        example: johndoe
        type: string
    type: object
...
  utils.ResponseDataAddedSocialMedia:
    properties:
      data:
        $ref: '#/definitions/utils.AddedSocialMedia'
      status:
        example: success
        type: string
    type: object
...
  utils.ResponseDataEditedSocialMedia:
    properties:
      data:
        $ref: '#/definitions/utils.EditedSocialMedia'
      status:
        example: success
        type: string
    type: object
  utils.ResponseDataEditedUser:
    properties:
      data:
        $ref: '#/definitions/utils.EditedUser'
      status:
        example: success
        type: string
    type: object
...
  utils.ResponseDataGetedSocialMedia:
    properties:
      data:
        $ref: '#/definitions/utils.SocialMedias'
      status:
        example: success
        type: string
    type: object
  utils.ResponseDataLoggedinUser:
    properties:
      data:
        $ref: '#/definitions/utils.LoggedinUser'
      status:
        example: success
        type: string
    type: object
  utils.ResponseDataRegisteredUser:
    properties:
      data:
        $ref: '#/definitions/utils.RegisteredUser'
      status:
        example: success
        type: string
    type: object
...
  utils.ResponseMessageDeletedSocialMedia:
    properties:
      message:
        example: your social media has been successfully deleted
        type: string
      status:
        example: success
        type: string
    type: object
  utils.ResponseMessageDeletedUser:
    properties:
      message:
        example: your account has been successfully deleted
        type: string
      status:
        example: success
        type: string
    type: object
  utils.SocialMedia:
    properties:
      created_at:
        example: here is the generated created at
        type: string
      id:
        example: here is the generated social media id
        type: string
      name:
        example: Example
        type: string
      social_media_url:
        example: https://www.example.com/johndoe
        type: string
      updated_at:
        example: here is the generated edited at
        type: string
      user:
        $ref: '#/definitions/mygram-byferdiansyah_socialmedia_utils.User'
      user_id:
        example: here is the generated user id
        type: string
    type: object
  utils.SocialMedias:
    properties:
      social_medias:
        items:
          $ref: '#/definitions/utils.SocialMedia'
        type: array
    type: object
host: localhost:8080
info:
  contact:
    email: ferdicompany@gmail.com
    name: ferdi
  description: MyGram is a free image sharing app written in Go. People can share,
    view, and comment images by everyone. Anyone can create an account by registering
    an email address and creating a username.
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  title: MyGram By Ferdiansya
  version: "1.0"
paths:
...
  /socialmedias:
    get:
      consumes:
      - application/json
      description: Get all social media with authentication user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseDataGetedSocialMedia'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_socialmedia_utils.ResponseMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_socialmedia_utils.ResponseMessage'
      security:
      - Bearer: []
      summary: Get all social media
      tags:
      - socialmedias
    post:
      consumes:
      - application/json
      description: Create and create a social media with authentication user
      parameters:
      - description: Add Social Media
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/utils.AddSocialMedia'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.ResponseDataAddedSocialMedia'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_socialmedia_utils.ResponseMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_socialmedia_utils.ResponseMessage'
      security:
      - Bearer: []
      summary: Add a social media
      tags:
      - socialmedias
  /socialmedias/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a social media by id with authentication user
      parameters:
      - description: SocialMedia ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseMessageDeletedSocialMedia'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_socialmedia_utils.ResponseMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_socialmedia_utils.ResponseMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_socialmedia_utils.ResponseMessage'
      security:
      - Bearer: []
      summary: Delete a social media
      tags:
      - socialmedias
    put:
      consumes:
      - application/json
      description: Edit a social media by id with authentication user
      parameters:
      - description: SocialMedia ID
        in: path
        name: id
        required: true
        type: string
      - description: Edit Social Media
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/utils.EditSocialMedia'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseDataEditedSocialMedia'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_socialmedia_utils.ResponseMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_socialmedia_utils.ResponseMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_socialmedia_utils.ResponseMessage'
      security:
      - Bearer: []
      summary: Edit a social media
      tags:
      - socialmedias
  /users:
    delete:
      consumes:
      - application/json
      description: Delete a user with authentication user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseMessageDeletedUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_user_utils.ResponseMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_user_utils.ResponseMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_user_utils.ResponseMessage'
      security:
      - Bearer: []
      summary: Delete a user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Edit a user with authentication user
      parameters:
      - description: Edit User
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/utils.EditUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseDataEditedUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_user_utils.ResponseMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_user_utils.ResponseMessage'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_user_utils.ResponseMessage'
      security:
      - Bearer: []
      summary: Edit a user
      tags:
      - users
  /users/login:
    post:
      consumes:
      - application/json
      description: Authentication a user and retrieve a token
      parameters:
      - description: Login User
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/utils.LoginUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseDataLoggedinUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_user_utils.ResponseMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_user_utils.ResponseMessage'
      summary: Login a user
      tags:
      - users
  /users/register:
    post:
      consumes:
      - application/json
      description: create and create a user
      parameters:
      - description: Register User
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/utils.RegisterUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.ResponseDataRegisteredUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_user_utils.ResponseMessage'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_user_utils.ResponseMessage'
      summary: Register a user
      tags:
      - users
securityDefinitions:
  Bearer:
    description: Description for what is this security definition being used
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"